[chap3] 챕터별 패키지 분리
